Getting rid of hittable class
