--- src/shapes/hittable.hpp
@@ -1,40 +1,0 @@
-#ifndef HITTABLE_H
-#define HITTABLE_H
-
-#include "core/ray.hpp"
-#include "core/utils.hpp"
-#include "core/vec3.hpp"
-
-namespace shkyera {
-
-class material;
-
-struct hitData {
-    point3 p;
-    vec3 normal;
-    shared_ptr<material> hitMaterial;
-    double t;
-    double u;
-    double v;
-    bool frontFace;
-
-    inline void setFaceNormal(const ray &r, const vec3 &outwardNormal) {
-        frontFace = dot(r.direction(), outwardNormal) < 0;
-        normal = frontFace ? outwardNormal : -outwardNormal;
-    }
-};
-
-class hittable {
-  public:
-    virtual bool hit(const ray &r, double minT, double maxT, hitData &data) const = 0;
-
-    virtual std::string getName() const = 0;
-    virtual void setName(std::string name) = 0;
-
-    virtual shared_ptr<material> getMaterial() const = 0;
-    virtual void setMaterial(shared_ptr<material> material) = 0;
-};
-
-} // namespace shkyera
-
-#endif
